Move screen thread work to screen.rs
